requirements.txt
mpi4py==3.0.0
numpy==1.15.0
h5py_cache==1.0
plotly==2.7.0
python_dateutil==2.7.3
ntpath==1534770254.427198
pysam==0.15.1
